Return correct value of the_excerpt
When searching the previous setup returned the_excerpt of the first post.
Because within the_excerpt(), $post was null and it returned the first post. Because it's not proper in "The_Loop"

We now pass $this->_post through `get_the_excerpt` and apply the_except filter if it gets filtered just in case.

When removing `setup_postdata()` it all stopped working... don't now yet why.