The file loaded by clarkson should never give a render warning
Sometimes a filter calls the clarkson render function before the plugin calls it. This results in the plugin throwing a warning when rendering the generic template file such as 'single.twig'. As this warning comes before any HTML content it sometimes messes up layouts and causes invalid documents when `WP_DEBUG` is `true`.