Fix finding the correct twig template file
Don't incorpate own custom logic because the third $templates parameter contains correct template name to search for according to the Theme Hierarchy.
This method, using the $templates array, also considers the egde-case / bug that @NielsdeBlaauw encoutered: Load the correct archive.twig template when you visit the archive page but there are not Posts in that archive. This because WordPress alraedy take this in account and we now really follow the WordPress flow by using $templates array.

Based on #99 of @christiaanstijnen and @Willemijnr. Thank you!