Rewrites Clarkson_Term creation.
Term creation was broken.

If a `get_term` returned `WP_Error`, the object would be created and fail at a later point.

You could arbitrarily set the taxonomy on a `\Clarkson_Term` and make it create an incorrect object.

We were retrieving complete `WP_Term` objects, just to find out the `term_id` and retrieve the `WP_Term` again in the constructor.