Correctly loads templates based on possibilities in WordPress.
Previously a template was an extension of a regular Clarkson_Object. This meant that a template could only extend one 'post_type'. For example, a `template_contact` could only extend a `page`.

However, in WordPress a template can be applicable to any number of post types. Basically, a `template_contact` could be set on a `page`, but also on a `custom_post_type`. In the old method, we would have to choose which object to extend, and maintain the other object to keep the same interface methods available.

In the new method the Clarkson Object gets loaded instead of a Clarkson_Template object. On the Clarkson_Object you can then ask `get_template()` which retrieves the template. Also, this allows object creation of standalone template objects, and adds `get_one` and `get_many` methods.

The `get_one` is usefull if you need a single instance of a template such as a contact page. You simple use `$contact_page = template_contact::get_one(); return $contact_page->get_required_fields() ?? array();`